requirements.txt
streamlit
easyocr
gspread
oauth2client
opencv-python-headless
numpy
Pillow